VM.create: Make readonly optional for disk devices
When processing disk devices to create a libvirt xml file, the current code
assumes that the 'readonly' attribute is always present.  If it's missing,
VM.create will fail with an AttributeError.  When describing disks, 'readonly'
is typically an optional parameter that defaults to False if not specified.  In
fact, this is what getConfDrives already does for legacy drives.  Right now,
there is only one place that assumes readonly is present.  Change it to check
for the attribute first.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Change-Id: I83b43ef02b38091a8b3bfd1a9d0485e3b04323c3
Reviewed-on: http://gerrit.ovirt.org/5298
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Bing Bu Cao <mars@linux.vnet.ibm.com>
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>