Adding the status field to getDeviceList info.
This status tries to predict the result of pvcreate on the lun.

By now the reported statuses are:
free: pvcreate --test on this lun succeeded.
used: pvcreate --test on this lun failed.

Caveat emptor: pvcreate on the device can fail even if "force"
flag is passed to createVG in case for example there is a mounted
file system on this lun.

Change-Id: I6af6628a29bb597cf46b57a2b439e1b950bef9ab
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/5515
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>