Move actual VM creation function to clientIF.py
To avoid using API in clientIF.py, we move the VM creation function
to clientIF.py. It should be safe to skip the VM parameters validation
in the VM.create() API for recovering exsiting VMs.

If this patch is accepted, we can add a singleton interface for
clientIF which can be used for API.py. Otherwise, it will cause circular
importing.

Change-Id: I38644263e4b4c3fe836a60405b251768db0a471c
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4808
Reviewed-by: Dan Kenigsberg <danken@redhat.com>