pthread.py: avoid stdout noise on process shutdown
For some reason, on process shutdown, pthread.PthreadMutex.__del__() is
sometimes called after the reference _libpthread to libpthread is reset
to None. This resutls in annoying message like
    AttributeError: 'NoneType' object has no attribute 'pthread_mutex_destory' (Ignored)

Current code catches this error and replaces it with a `print`
statement, which is just as annoying.

This patch intentionally swallows the ignored error (and only it)

Change-Id: Ib2fc9ca3b258d99c6ce56069f908b09cdfd6d61a
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/5659
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Tested-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>