BZ#835784 - Allow to create a network on top of existing bond in additional to create a new bond and network
If we already have a bond0 with (eth1, eth2) and now we want to create
network "bridge_woVlan" on top of it and in additional we want to create
a new bond1 with (eth3, eth4) and create a network "bridge_woVlan_2" on
top of bond 1.
This patch will allow us to do it in the same setupNetwork command.

Change-Id: I67d24d656934ea9dcb0a8209c6a375e19c23f82b
Signed-off-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/5737
Reviewed-on: http://gerrit.ovirt.org/5759
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>