Added tests for _addNetworkValidation and helpers.
Helpers are: validateVlanId, validateBondingName and validateBondingOptions.

The test for validateBondingOptions monkey patches os.path.exists to allow to
test just the logic of the method, which uses sysfs to determine if the option
is valid.

The test of _addNetworkValidation goes over the logic branches to test that
the proper exceptions are thrown. It monkey patches the network detection
method of the netinfo module to avoid real network detection at test time.

Change-Id: Ifd8bb649685fac00640e894a4484cdb0cc0b2f1c
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/6198
Tested-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>