configNetwork: fix bonding mode issue
Currently, we configure nics before bonding when create/edit bonding
device. That could cause the bonding mode can't be updated as specified.
Because, when we use ifup to activate the bonding slaves, it implicitly
creates the bonding device with the default mode 0, and adds the salve
nics to it. So you will see the following kernel message in system log:

bonding: unable to update mode of bondX because it has slaves

The fix is just to make bonding configured before nics. Then it can
update its mode because no slave attatches to it at that time.

Change-Id: I9855ef2d0e05603f07397eb26822fac5fc8eaf5b
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/6217
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>