vdsm_api: Add schema and html doc generator
This patch adds a schema file to a new directory 'vdsm_api'.  The schema is
intended to be a machine-readable specification of the current vdsm API.  From
this schema we should be able generate API documentation and a client library
(libvdsm) that other projects can use to write against the defined API.

At this point, the schema file probably contains some errors and needs a solid
review from this community.  Please send patches to gerrit that correct
documentation and data errors in the schema.  Meanwhile, I will begin to write
a gObject code generator.

Change-Id: I703703ba212b7560d1054f32a86ad083ee88c6c8
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/6343
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>