Add the hostId parameter to reconstructMaster
During reconstructMaster a new storage domain of the pool is elected
to become the master to hold the cluster lock and the metadata.
With the introduction of SANLock each host maintains its own host id
on all the storage domains of the pool and therefore VDSM can't use a
temporary id to reconstruct the master domain.

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I5c85d9440b467eb755d196a6548d2f7b12bcab9f
Reviewed-on: http://gerrit.ovirt.org/6562
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>