fix 'qemuImg' import error in unit test
restTest indirectly import qemuImg.py. qemuImg.py is installed to
vdsm site-packages dir, so it is imported as "from vdsm import qemuImg"
in other files.
There is a hack in testrunner.py to import modules under vdsm
site-packages dir for unit test, but qemuImg is not hacked, so causes the
problem.

This patch add qemuImg to the import hack list in testrunner.py and fix
the problem.

Change-Id: I7b958b1aa6ad8d912bd3f191478ed181fbd14bb4
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/6643
Reviewed-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Tested-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>