BZ#844656 Respawn the domain monitor when needed
In the domain monitor there are few lines that are not covered by a
try/except clause but that might be enough for the thread to die without
leaving any valuable information in the logs. This patch adds logging
for the possible exceptions and respawn the domain monitor if it exits
unexpectedly.

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I47e4fb6883d72578ee8083e46e4f660b656a5ffc
Reviewed-on: http://gerrit.ovirt.org/6787
Tested-by: Omri Hochman <hoomri@gmail.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>