Related to BZ#844347 - Keep bond mode after change network on top of it
Assume we have a bond with mode X and network on it. Now we want to remove this network
and add another one.
In this case setupNetwork will get a new network but not a bond info.
We need to take bond info from the host and keep it during network addition

Change-Id: Idac51ecbef9d11f31e9bd962d3aefbde34e3fb54
Signed-off-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/6805
Reviewed-by: Dan Kenigsberg <danken@redhat.com>