report network's underlying device as "iface"
ovirt-engine-3.1 expects to see "interface" for bridgeless networks, and
only for them. A recent patch http://gerrit.ovirt.org/6644 started
reporting "interface" for bridged networks, too. Which broke Engine.

I consider this an Engine bug (it should have relied on the "bridged"
attribute to decide if a network is bridged or not), but this cannot be
changed on the Engine side so near to the ovirt-3.1 release.

This patch reports "iface" for all networks, keeping "interface" only
for bridgeless networks, for ovirt-3.1 backward compatibility.

With this new API nit, we report a new cluster level 3.2.

Change-Id: Ibc8a5b8c66ecbb5cfd6378db358e38375b16c54c
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/6806
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>