vdsmd: Don't keep the temporary file generated by configuring libvirt
It turns out that this problem is caused by a sed bug. And it has been
fixed in https://bugzilla.redhat.com/show_bug.cgi?id=832855. This patch
bumps the required version of sed to 4.2.1-10 to include the fix. The
sed bug doesn't exist on rhel6, so we don't need to change it.

Change-Id: Ic5c0fe8b86c4fa82bb89d581f594b0dfb9aae425
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/4864
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>