BZ#846604 bootstrap: skip clock sync if ntpd is synced
During boostrap, host clock is synchronised with
engine time. If on host ntpd is synced and on engine
time is not synced, time in host will jump to engine
time then jump back to ntp time few moments later.
This patch checks if ntpd is synced in order to
skip host clock sync with engine time when true.

Change-Id: Ie594dfef430c811f920f304c5182be9e34f6fbf1
Signed-off-by: Amador Pahim <apahim@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7070
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>