bootstrap: introduce bootstrap interface version
Bootstrap interface version is a method of supporting multiple engine
versions within bootstrap code.

The vdsm-bootstrap code is separate package which can be upgraded
without engine upgrade. Engine is also upgrade protected, so force
update upon vdsm-bootstrap upgrade is difficult.

As the vdsm-bootstrap is separate package from engine, but depends
heavily on engine interaction, and since it may be updated, it should
support backward compatibility with different versions of engines.

If changes are relatively small, a bootstrap interface version may come
in handy so no script duplication is required.

Change-Id: I67b92dce5896a5cf510ffc0d51450e56070eb03e
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7092
Reviewed-by: Dan Kenigsberg <danken@redhat.com>