BZ#840594 Fix netinfo when libvirt bridge does not exist.
In the case where a bridge that is configured in libvirt is manually removed or
renamed, netinfo throws an exception when trying to discover which ports are
assigned to the bridge.

This patch catches the OSError exception of trying to list the ports of a non-
existing bridge interface, effectively filtering the non-existing interface
from the network capabilities.

Also adds a unittest for getting network information of non-existing bridges.

Change-Id: I2538a8ce2fafd8efed834b87b20a63d058afd96b
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7127
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>