schema: Remove invalid names
The identifier "type" causes conflicts with the GObject type system.  Type is
not a great name anyway so choose better names where it occurs.

The name "StorageDomainClass" conflicts with the GObject class structure for the
StorageDomain class.  Rename it to StorageDomainImageClass since it identifies
the class of images stored on the storage domain.

Change-Id: Idb43f5960f60a2064a5a41eeafcb7a7ef1899c77
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7013
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Tested-by: Saggi Mizrahi <smizrahi@redhat.com>