Fix a race existing in the singleton interface of clientIF
There's a race in getInstance() between clientIF initialization and MOM thread.
If for some reason the initialization of clientIF is slowed down and not finished
before MOM calls vdsm API, an exception will happen in MOM thread due to that
the parameter 'log' is missing. The fix is just moving the parameter check into
the protection of the singleton lock. Then MOM thread will be blocked if it tries
to access vdsm API when the initialization of clientIF is ongoing.

Change-Id: I6d907f5613963438c99749da674bb6588a23fe38
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7283
Reviewed-by: Royce Lv <lvroyce@linux.vnet.ibm.com>
Reviewed-by: Adam Litke <agl@us.ibm.com>
Tested-by: Adam Litke <agl@us.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>