improve message when trying to attach import domain with wrong permissions
When a storage domain is attached, and there is something wrong with the
directory permissions, then the operation fails with "Cannot obtain
lock" error. This patch improves the error in this case, by performing
validations when doing the server connection. If the permission
validation fails, we disconnect.

The user will get an event specifying the permissions error.
And, he will get a UI error saying:
"There is no storage domain under the specified path. Please check
path."

This fix also applies to posix and local storage.

Bug-Id: BZ#842146
Change-Id: I76880b1445f259431d7aa691eda676e1210308e4
Signed-off-by: Oved Ourfali <oourfali@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7339
Reviewed-by: Eduardo <ewarszaw@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>