confNet: ifcfg's VLAN should be set only on vlan cfg
If a misguided script calls addNetwork with explicit VLAN option, it
should not be blindly copied to the top-most device.
We set VLAN=yes on the ifcfg script for the vlan device, and that's
that.

This is a very limited hack of a wider problem: unrecognized network
options are copied to the bridge's ifcfg file.

Bug-Url: https://bugzilla.redhat.com/1051150
Change-Id: Ia2c9d6c86e73104424ca8ac901bf09e4b98a47ab
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7411
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24842
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>