Add an option to create a watchdog device.
A support for a watchdog device was added to the "devices"
configuration.

Also add the watchdog event callback to log the watchdog action
of VM.

model of watchdog:
    'i6300esb' default model, emulating a PCI Intel 6300ESB
    'ib700'    emulating an ISA iBase IB700
Only one watchdog device is supported for each VM.

action of watchdog timeout:
    'reset'     forcefully reset the guest
    'shutdown'  gracefully shutdown the guest (not recommended)
    'poweroff'  forcefully power off the guest
    'pause'     pause the guest
    'none'      default, do nothing
    'dump'      automatically dump the guest

the parameter of 'watchdog' device as follow:
{'device': 'watchdog', 'type': 'watchdog',
 'specParams': {'model': 'i6300esb', 'action': 'none'}}

The watchdog device can be used to detect guest crash or hang, and
if 'dump' is chosen for the action of watchdog timeout, libvirt will
dump guest's memory on timeout automatically.

The directory to save dump files can be configured by auto_dump_path
in file /etc/libvirt/qemu.conf.

The watchdog device requires an additional driver and management
daemon in the guest. Just enabling the watchdog in the vdsm
"devices" configuration does not do anything useful on its own.

Change-Id: I2b7970a9050ab0279fe03371b9a77692fba30af8
Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7535
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>