bootstrap: node id: consider bonding MAC addresses as well
The node id is constructed using the lowest MAC address.

If bonding is applied after initial node id was taken, it may shade the
minimal address.

This change adds to the MAC addresses of bonding interfaces to the MAC
address addresses to consider.

Sync with engine command used for same purpose.

Change-Id: Iae0713f95c25e3e05b24bf8254499e4cd8e64d10
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7596
Reviewed-by: Barak Azulay <bazulay@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>