tc: avoid adding non-existing devices to filter
When a mirroring target dies, its tap device are removed. On `tc filter
show`, the tap device name is replaced with a string like `if17`, which
is not a name of an existing device.

This patch is somewhat racy, as another qemu may die just before the new
filter is set.

Change-Id: I9c1b459557985a3a894f182880411889f2c8a06f
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7631
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Tested-by: Meni Yakove <myakove@redhat.com>