vdsm: Unifying engine and vdsm commit templates
As agreed in ovirt's arch mailing-list[1], we'd like the engine's
commit template and vdsm conventions to agree on BZ referencing.

This change adds a sample commit message template with a 'Bug-Url:',
which matches the mentioned agreement. The way to use this sample
is with: 'git commit -s -F vdsm-commit-template.txt -e'. Usage of
the template is optional, and meant for convenience only.

[1] http://lists.ovirt.org/pipermail/arch/2012-September/000838.html

Change-Id: I9b8769fbf65ceab391443a4fbe2eb0fcdec5d6f4
Signed-off-by: Doron Fediuck <dfediuck@redhat.com>
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7699
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>