remove the broken support for delivering video to local display
In vdsClient help message and sample VM configurations, it is said when
display is set to local, the VM video will go to local host. In fact
vdsm do not implement this. If the user use this feature, vdsm will
raise an exception in function getConfVideo() in vdsm/vm.py.

To implement this feature, we have to
1. deal with 'local' properly in getConfVideo() in vdsm/vm.py, add a
suitable video device when 'local' is set.
2. generate graphic discription XML node properly in vdsm libvirtvm.py,
like <graphics type="sdl" display=":0" xauth="blah">. Now we always
generate some attributes not related to SDL, and libvirt will complain.
We have to suppress generating those attributes when using local
display.
3. If we allocate a host display for the user qemu, then qemu can use sdl
to show the display in its own host display.
4. If we do not want to allocate a display for qemu, we have to use an
existing one, then we have to deal with the $DISPLAY(:0 ?) and
$XAUTHORITY. These two attributes can be set in the libvirt XML in the
graphics section
5. We often setup hosts without X support, it is not possible to
deliver the guest video to host display. In this situation, detect the
host X support and raises an exception properly to the user to describe
what's wrong.

If we are not going to do the work as written above, we'd better remove
the broken support for local display.

Change-Id: I4e655b7e03be74c5d9627ec03ac2baefcc379681
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7725
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>