fileSD.getDomUuidFromMetafilePath: return correct sdUUID
Since commit 7ce1a3d53, getDomUuidFromMetafilePath() returns a constant
string "dom_md", regardless of the storage domain metafile it receives.

This has serious effects, since it leads to all NFS domains using a
single process pool.

It would probably be better to solve the problem by avoiding
getDomUuidFromMetafilePath altogether. The callers high up the stack
know the sdUUID they are working on. This solution is left as an
excersice to the readers.

A simpler, localized, solution is to take the correct element out of
metafile path. In this patch I suggest to do it without worrying about
trailing slashes in the [irs]repository configurable.

Thanks to Shu Ming for reporting this bug and suggesting another fix for
it.

Change-Id: I9931a1aa804d9dc7aec1485ebc29bac45131d0e7
Signed-off-by: Shu Ming <shuming@linux.vnet.ibm.com>
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7078
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>