glusterHostsList verb returns local hostname as IP address or FQDN
The verb glusterHostsList executes the 'gluster peer status' command
to fetch list of servers of the gluster cluster.  However the output
of this command doesn't include the server on which the command is
being executed.  For completeness the verb appends current server
detail to the list given by gluster

When hostname of the current server is used in the detail, the problem
with hostname is that what's saved in engine db in the host_name field
is not necessarily the actual hostname of the host, but whatever user
entered while adding the host, which could be ip address also.  So
engine can't rely on the hostname, particularly on compare the list
returned by the glusterHostsList verb with the list of hosts present
in engine DB to identify if something has changed because of direct
cli operations from the gluster cli.

So this patch enables sending an ip address instead of hostname of the
current server.  Engine db contains all ip addresses of the server,
and hence ip address can be used for comparisons.

Change-Id: I24d2b4d27107d208672b1d3df96dd7759d2288c5
Signed-off-by: Bala.FA <barumuga@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7857
Reviewed-by: Shireesh Anjal <sanjal@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>