faqemu: remove the fake /dev/kvm after vdsmd stops
/usr/libexec/vdsm/vdsm-faqemu, the fake emulator, will create a fake
/dev/kvm before it starts a real emulator. This fake device is for
cheating libvirtd. This fake /dev/kvm is not deleted automatically after
vdsmd stops. If we want to switch to real KVM support then, the fake
device will prevent the system creating the real device, thus makes the
user very confused about the error. This patch detects and deletes the
fake /dev/kvm after vdsmd stops whenever faqemu is enabled or not in the
vdsmd.conf.

Change-Id: I1c1183acf90cd32444b1c26fa20aa843e4a0556d
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7885
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>