image: avoid prezeroing optimization in some cases
The prezeroing optimization should be used only when it is actually
stricly required. Using such optimization with COW is dangerous since
the qcow2 header would contain the temporary volume size (10Mb).

Until now this wasn't a problem because the content of the volume
(qcow2 header included) would have been replaced by the original
content (during a copy), with the introduction of live storage
migration this might become a problem because the leaf is actually
used for mirroing.

Change-Id: I9b8f348af355b0f17ab66dd9d5ef57c905051c17
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7935
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>