vdsm-reg-setup: get hostname with gethostname()
Previously we relied upon the fallback behavior of socket.getfqdn()[1] which returned
the hostname only in the case that it could not find an entry or alias with a dot.
In ovirt-node if the user sets a new hostname, this adds an alias to /etc/hosts
that maps to localhost.local domain and leads to engine not listing the
user-requested hostname. The fix is to not use getfqdn(), but gethostname().

Test
=====
* Install oVirt Node 2.5.1-1.0
* Set Network (default it will be localhost.localdomain)
* Select Network again, and change Hostname (supernode.linux.com)
* Register the Node into Engine
- At this moment, Engine will show localhost.localdomain instead of supernode.linux.com

The manual workaround is remove the new hostname from /etc/hosts in 127.0.0.1 entry.

[1] http://docs.python.org/dev/library/socket.html

Change-Id: I38ec4c570d7f250d64e62e01b91cd2828780c45b
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/7939
Reviewed-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-by: Barak Azulay <bazulay@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>