Validate downloaded SSH public key
During registration the SSH public key of the engine is downloaded,
but it is not verified in any way. In some situations we can be
getting garbage from the server, for example when we connect to an
HTTPS server using the HTTP protocol. That garbage can end up in the
uthorized_keys file.  This patch changes deployUtils.py so that it
verifies that what we get looks like a valid SSH public key before
saving it to the file.

Bug-Id: https://bugzilla.redhat.com/856167
Change-Id: Ic6873690534f431867859e91b5fb6b1f693155aa
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8018
Reviewed-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>