BZ#856167 - Parse correctly the management_server parameter
When doing automatic installations we parse the management_server
parameter to find the engine address and port number, but we don't
parse it correctly if the port number is missing, in this case we end
up using the address as both address and port number. This patch fixes
that so when the port number is missing the default 443 will be used.

Change-Id: I74d658902d3421c0ac554c5c2634ddc8c6a28b21
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8022
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>