Validate downloaded CA certificate
During installation the CA certificate of the engine is downloaded,
but it is not verified in any way. In some situations we can be
getting garbage from the engine, for example when we connect to an
HTTPS server using the HTTP protocol. That garbage can corrupt a
previously downloaded CA certificate. This patch changes
deployUtils.py so that it verifies that what we get is a valid X.509
certificate before saving it to the file.

Bug-Id: https://bugzilla.redhat.com/856167
Change-Id: Ib5d3b3aeca42e4bc4b621b1acb861bfb1ac383e6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8021
Reviewed-by: Ryan Harper <ryanh@us.ibm.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>