Fix data passing to process when doing async execution
There is a mistake in the current file where the stream data is being
sent to when using `sync=False` is stdout instead of stdin. This makes
the feature not work.

Further more, I added flush so that the data is sent before the function
returns because it's a bit more intuitive.

Also, add a test to check that this feature works so in the future this
kind of thing wouldn't occur.

Change-Id: Ibe496e74a24721ba00c596d36286bcdafdc20069
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8081
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Shu Ming <shuming@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>