vdsm.spec.in: use locale LC_ALL=C
During VDSM instalation, we verify if SELinux booleans are
enabled with return of semanage command, output like: "on" or "off".
These are English words and if the system contains a different
locale, example pt_BR the validation will fail.

Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=857556

This patch will change the locale for English (C) while
installation VDSM.

Change-Id: I22b5706e02c71f463200ce20eb0e426e1bfee6e6
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8068
Reviewed-by: Dan Kenigsberg <danken@redhat.com>