storage: Change low-priority tasks to idle priority
Some tasks run as best-effort priority level 7, which is not low enough
to prevent starvation of higher-priority tasks on the host.  This can
hurt response time and even cause timeouts under heavy load.

Instead, use idle priority so that the operations are done without
interrupting other system operations.

Note that idle priority behavior changed in kernel 2.6.25; there is now
a guarantee of forward progress for processes at this level, making it
usable for tasks such as these.

Bug-Url: https://bugzilla.redhat.com/854242
Change-Id: I1b3a9088ad550c6d235d22d00688673c2cac0f7e
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8121
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>