Fix duplicate emulated machines returned
On Fedora, libvirt call getCapabilities() also append the list of emulated
machines to the 'domain' node whose type is 'kvm'. The list is identical to
what's listed in the default info of that arch. So it would cause
getElementsByTagName('machine') matches all of machine tags and return a list
of duplidate emuated machines.

This patch changes to only collect information from the direct children of
tag 'arch' and adds a test case for it.

Change-Id: I7e28688bf09b5bc413afffe9cc1501d4247d67e0
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/8208
Reviewed-by: Hunt Xu <mhuntxu@gmail.com>
Reviewed-by: Adam Litke <agl@us.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>