Apply memoized to getos()
getos() is called by _getKeyPackages() every time the capabilies
are queried. For the case of RHEL system, it has to call
os.path.exists() four times to get the OS name.  So it's worth to
memory it by the decorator memoized. And it will not change while
vdsm is running, so it's safe to be cached.

Change-Id: Id878b558ff646334939137d23fb05bd065e0e9d6
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/8209
Reviewed-by: Dan Kenigsberg <danken@redhat.com>