Store engine web CA cert in engine_web_ca.pem
Currently we store the CA certificate downloaded from the engine for
registration purposes in the /etc/pki/vdsm/cacert.pem file.  This file
is then replaced by VDSM by its default one during reboot, making a
backup before. This means that after the reboot vdsm-reg can't use it to
download the SSH key, and this means that registration fails.

This patch changes deployUtil.py so that it downloadto s the certificate
of the CA that signs the certificate of the engine web server to a new
file: /etc/pki/vdsm/enginecacert.pem. This file is not touched by the
VDSM start script, so that vdsm-reg can use it later to download the SSH
key correctly.

Change-Id: I127bf44cbcde90f7dae26a3bd3127f3eac2ca53c
Bug-Url: https://bugzilla.redhat.com/856167
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8038
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>