libvirtconnection: retry opening a libvirt connection
When we try to open a connection to libvirt too early after libvirtd has
been started, we fail. This was first noticed in vdsmd after reboot, and
was hacked there. However, the same bug appears when vdsm-reg-setup is
run after reboot, and may appear during bootstrap.

This patch polls for 10 seconds when a new connection to libvirt is
requested.

Bug-Url: https://bugzilla.redhat.com/862002
Change-Id: Ia3792281d8f092da46861cacc4a4514e96fc842b
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8372
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>