bootstrap: allow customize bridge name for standard sequence
For gluster integration, we need bootstrap to be able to accept
management bridge name within bootstrap and not relay on vdsm build time
settings.

As gluster only performs standard host bootstrap (as opposed to
registration), the customization will be available only for this
sequence.

In future the registration sequence will be altered to match the standard
sequence.

Change-Id: I70c178de55e14ecdcff92270df489dc1415ef26f
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=863717
Reviewed-on: http://gerrit.ovirt.org/8388
Reviewed-by: Dan Kenigsberg <danken@redhat.com>