Don't fail silently when ifup fails.
Up until now we discarded ifup return codes and ignored the possible
errors present in ifup stdout. Use this information to perform better
error reporting.

The tests change is due to the fact that now that we treat ifup
errors as excepcional occurences, for unit testing the persistent
backups (thing which can be done by a regular user) we do not need
to test ifup and ifdown of the current config as it is not a thing
that the unprivileged user is normally allowed to do.

Change-Id: I1cc9dcc0a6b55d36fc937e1d364bd9c256ecd70a
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Bug-Url: https://bugzilla.redhat.com/856737
Reviewed-on: http://gerrit.ovirt.org/8415
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>