volume: remove domain shared lock from getVolumeSize
During long operations involving an exclusive lock on a storage
domain (e.g.: domain upgrade) the getVolumeSize calls coming from
the VmStatsThread are queued and the statsAge grows until the VMs
are moved to not-responding.

Also the extendVolume requests are queued (with requests coming
from other hosts in the cluster) and the (redundant) refreshVolume
issued by the VMs on the SPM.

Removing the getSharedLock call from such methods is the safest
way to fix the issue since we don't use more granular locks during
the upgrade (image/volume locks) and these calls can either succeed
or fail without undesirable side effects.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=865386
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I644fff245b62ef009a4ca7b6f4a903c9ffcc9653
Reviewed-on: http://gerrit.ovirt.org/8517
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>