storage: failure to reconnect to authenticated iSCSI LUNs
Fix failure to reconnect to iSCSI storage domains on LUNs requiring
authentication due to missing CHAP credentials.

The credentials were sent to VDSM when connecting sessions on the host,
but VDSM was not sending them back to the engine when the engine queried
for connection properties.  Without this, the engine couldn't store them
in the database, and thus couldn't send them back out to any host trying
to connect (or reconnect) the storage.

Change-Id: Id717ee5b45adede4f9094f842ead06b0ba91f8b2
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8523
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>