schema: add explanation string for the ballonInfo field
Recently http://gerrit.ovirt.org/#/c/3669/ is merged. It adds a new
member named balloonInfo to the type RunningVmStats, but forgets to add
the corresponding information string in the comments. This will cause
process-schema.py raise a KeyError when looking for info_data of
balloonInfo and fail to generate the html documentation.

This patch adds the corresponding information string.

Change-Id: I3db7251a0451c48143d1fcd8ad8d8ade242ac372
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/8599
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>